Making name plural, and adding support for any  user
